# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Web Scanner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ScanScanType = typeof ScanScanType[keyof typeof ScanScanType];


export const ScanScanType = {
  document: 'document',
  photo: 'photo',
} as const;

export type ScanColorMode = typeof ScanColorMode[keyof typeof ScanColorMode];


export const ScanColorMode = {
  color: 'color',
  greyscale: 'greyscale',
} as const;

export type ScanFormat = typeof ScanFormat[keyof typeof ScanFormat];


export const ScanFormat = {
  pdf: 'pdf',
  jpeg: 'jpeg',
} as const;

export interface Scan {
  id: string;
  name: string;
  pageCount: number;
  scanType: ScanScanType;
  colorMode: ScanColorMode;
  paperSize: string;
  format: ScanFormat;
  /** @nullable */
  thumbnailUrl: string | null;
  createdAt: string;
}

export type ScanInputScanType = typeof ScanInputScanType[keyof typeof ScanInputScanType];


export const ScanInputScanType = {
  document: 'document',
  photo: 'photo',
} as const;

export type ScanInputColorMode = typeof ScanInputColorMode[keyof typeof ScanInputColorMode];


export const ScanInputColorMode = {
  color: 'color',
  greyscale: 'greyscale',
} as const;

export type ScanInputFormat = typeof ScanInputFormat[keyof typeof ScanInputFormat];


export const ScanInputFormat = {
  pdf: 'pdf',
  jpeg: 'jpeg',
} as const;

export interface ScanInput {
  name: string;
  pageCount: number;
  scanType: ScanInputScanType;
  colorMode: ScanInputColorMode;
  paperSize: string;
  format: ScanInputFormat;
  /** @nullable */
  thumbnailUrl?: string | null;
}

export interface EmailShareInput {
  to: string;
  subject: string;
  message?: string;
  fileName: string;
  fileBase64: string;
  mimeType: string;
}

export interface EmailShareResult {
  success: boolean;
  message: string;
}