# lib/api-zod/src/generated/types/emailShareInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Web Scanner API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailShareInput {
  to: string;
  subject: string;
  message?: string;
  fileName: string;
  fileBase64: string;
  mimeType: string;
}