# lib/api-zod/src/generated/types/emailShareResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Web Scanner API
 * OpenAPI spec version: 0.1.0
 */

export interface EmailShareResult {
  success: boolean;
  message: string;
}