# lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Web Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanColorMode } from './scanColorMode';
import type { ScanFormat } from './scanFormat';
import type { ScanScanType } from './scanScanType';

export interface Scan {
  id: string;
  name: string;
  pageCount: number;
  scanType: ScanScanType;
  colorMode: ScanColorMode;
  paperSize: string;
  format: ScanFormat;
  /** @nullable */
  thumbnailUrl: string | null;
  createdAt: Date;
}