# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Web Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputColorMode } from './scanInputColorMode';
import type { ScanInputFormat } from './scanInputFormat';
import type { ScanInputScanType } from './scanInputScanType';

export interface ScanInput {
  name: string;
  pageCount: number;
  scanType: ScanInputScanType;
  colorMode: ScanInputColorMode;
  paperSize: string;
  format: ScanInputFormat;
  /** @nullable */
  thumbnailUrl?: string | null;
}